process: rename terminalsource to terminal_source
